Extracted template from film-card component
